pyproject.toml
[project]
name = "data-ingest"
version = "0.1.0"
description = "Document processing pipeline with contextual chunking, semantic embeddings, and vector database indexing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.41.0",
    "datasets>=4.4.1",
    "numpy>=2.0.0",
    "pinecone>=5.0.0",
    "python-dotenv>=1.2.1",
    "structured>=0.1",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "unstructured-client>=0.42.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
